wip: refactor insertion port
